fix: migrate to int64

+++ pkg/cluster/controller/groups/variables/zz_controller_test.go
@@ -42,7 +42,7 @@ import (
 
 var (
 	errBoom             = errors.New("boom")
-	groupID             = 5678
+	groupID             = int64(5678)
 	variableKey         = "VARIABLE_KEY"
 	variableValue       = "1234"
 	variableType        = commonv1alpha1.VariableTypeEnvVar
